Revert with segment binder in pr#34384
